fix: optimize modal in create new clan, check noti exist in list to group on android
