+++ apps/mobile/ios/Mobile/Info.plist
@@ -74,6 +74,10 @@
 	</dict>
 	<key>NSAppleMusicUsageDescription</key>
 	<string>We need access to Apple Music to play your selected songs</string>
+	<key>NSBluetoothAlwaysUsageDescription</key>
+	<string>We need needs to use Bluetooth to detect connected headsets for mezon meeting</string>
+	<key>NSBluetoothPeripheralUsageDescription</key>
+	<string>We need needs to use Bluetooth to detect connected headsets for mezon meeting</string>
 	<key>NSCameraUsageDescription</key>
 	<string>Mezon app needs access to the camera to call video or take add pricture send to chat</string>
 	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
